Reworked permission granting
