+++ app/src/main/java/com/tari/android/wallet/ui/fragment/contact_book/root/ContactBookFragment.kt
@@ -18,8 +18,6 @@ import com.tari.android.wallet.databinding.FragmentContactBookRootBinding
 import com.tari.android.wallet.extension.observe
 import com.tari.android.wallet.ui.common.CommonFragment
 import com.tari.android.wallet.ui.component.tari.toolbar.TariToolbarActionArg
-import com.tari.android.wallet.ui.extension.PermissionExtensions.runWithPermission
-import com.tari.android.wallet.ui.extension.PermissionExtensions.runWithPermissions
 import com.tari.android.wallet.ui.extension.setVisible
 import com.tari.android.wallet.ui.extension.string
 import com.tari.android.wallet.ui.fragment.contact_book.contacts.ContactsFragment
@@ -42,23 +40,21 @@ class ContactBookFragment : CommonFragment<FragmentContactBookRootBinding, Conta
         bindViewModel(viewModel)
 
         subscribeVM(viewModel.shareViewModel)
-        subscribeVM(viewModel.shareViewModel.bluetoothAdapter)
+        subscribeVM(viewModel.shareViewModel.tariBluetoothServer)
+        subscribeVM(viewModel.shareViewModel.tariBluetoothClient)
         subscribeVM(viewModel.shareViewModel.deeplinkViewModel)
 
-        viewModel.shareViewModel.bluetoothAdapter.init(this)
-
-        startReceiving()
+        viewModel.shareViewModel.tariBluetoothServer.init(this)
+        viewModel.shareViewModel.tariBluetoothClient.init(this)
 
         setupUI()
 
         subscribeUI()
 
-//        initTests()
-    }
-
-    override fun onResume() {
-        super.onResume()
         grantPermission()
+        startReceiving()
+
+//        initTests()
     }
 
     private fun subscribeUI() = with(viewModel) {
@@ -75,18 +71,19 @@ class ContactBookFragment : CommonFragment<FragmentContactBookRootBinding, Conta
 
         observe(shareViewModel.shareText) { shareViaText(it) }
 
-        observe(shareViewModel.launchPermissionCheck) { runWithPermissions(*it.toTypedArray(), openSettings = true) {
+        observe(shareViewModel.launchPermissionCheck) { permissionManagerUI.runWithPermissions(*it.toTypedArray(), openSettings = true) {
             viewModel.shareViewModel.startBLESharing()
         } }
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
-        viewModel.shareViewModel.bluetoothAdapter.handleActivityResult(requestCode, resultCode, data)
+        viewModel.shareViewModel.tariBluetoothServer.handleActivityResult(requestCode, resultCode, data)
+        viewModel.shareViewModel.tariBluetoothClient.handleActivityResult(requestCode, resultCode, data)
     }
 
     private fun grantPermission() {
-        runWithPermission(android.Manifest.permission.READ_CONTACTS, false) {
+        permissionManagerUI.runWithPermission(android.Manifest.permission.READ_CONTACTS, false) {
             viewModel.contactsRepository.contactPermission.value = true
             viewModel.contactsRepository.phoneBookRepositoryBridge.loadFromPhoneBook()
         }
@@ -153,9 +150,9 @@ class ContactBookFragment : CommonFragment<FragmentContactBookRootBinding, Conta
     }
 
     private fun startReceiving() {
-        val permissions = (viewModel.shareViewModel.bluetoothAdapter.bluetoothPermissions + viewModel.shareViewModel.bluetoothAdapter.locationPermission)
-        runWithPermissions(*permissions.toTypedArray(), openSettings = false) {
-            viewModel.shareViewModel.bluetoothAdapter.startReceiving()
+        val permissions = (viewModel.shareViewModel.tariBluetoothServer.bluetoothPermissions + viewModel.shareViewModel.tariBluetoothServer.locationPermission)
+        permissionManagerUI.runWithPermissions(*permissions.toTypedArray(), openSettings = false) {
+            viewModel.shareViewModel.tariBluetoothServer.startReceiving()
         }
     }
