+++ app/src/main/java/com/tari/android/wallet/ui/fragment/restore/enterRestorationPassword/EnterRestorationPasswordViewModel.kt
@@ -106,6 +106,6 @@ class EnterRestorationPasswordViewModel : CommonViewModel() {
             cancelable = false,
             canceledOnTouchOutside = false,
             onClose = { _backPressed.call() })
-        _modularDialog.postValue(args.getModular(resourceManager))
+        modularDialog.postValue(args.getModular(resourceManager))
     }
 }
