+++ app/src/main/java/com/tari/android/wallet/ui/fragment/send/addAmount/AddAmountViewModel.kt
@@ -109,7 +109,7 @@ class AddAmountViewModel : CommonViewModel() {
                 ButtonModule(resourceManager.getString(R.string.common_cancel), ButtonStyle.Close)
             )
         )
-        _modularDialog.postValue(args)
+        modularDialog.postValue(args)
     }
 
     fun calculateFee(amount: MicroTari, walletError: WalletError) {
