+++ app/src/main/java/com/tari/android/wallet/ui/fragment/settings/baseNodeConfig/addBaseNode/AddCustomBaseNodeViewModel.kt
@@ -77,7 +77,7 @@ class AddCustomBaseNodeViewModel : CommonViewModel() {
     fun onAddressChanged(text: String) = _onionAddressText.postValue(text)
 
     private fun addBaseNodePeerFailed() {
-        _modularDialog.postValue(
+        modularDialog.postValue(
             ErrorDialogArgs(
                 resourceManager.getString(R.string.common_error_title),
                 resourceManager.getString(R.string.debug_edit_base_node_failed)
