+++ app/src/main/java/com/tari/android/wallet/ui/fragment/settings/deleteWallet/DeleteWalletViewModel.kt
@@ -35,6 +35,6 @@ class DeleteWalletViewModel : CommonViewModel() {
                 ButtonModule(resourceManager.getString(R.string.common_cancel), ButtonStyle.Close)
             )
         )
-        _modularDialog.postValue(args)
+        modularDialog.postValue(args)
     }
 }
