+++ app/src/main/java/com/tari/android/wallet/ui/fragment/settings/networkSelection/NetworkSelectionViewModel.kt
@@ -56,7 +56,7 @@ class NetworkSelectionViewModel : CommonViewModel() {
                 resourceManager.getString(R.string.all_settings_select_network_confirm_description),
                 onConfirm = { changeNetwork(networkViewHolderItem.network) }
             )
-            _modularDialog.postValue(confirmDialogArgs.getModular(resourceManager))
+            modularDialog.postValue(confirmDialogArgs.getModular(resourceManager))
         } else {
             changeNetwork(networkViewHolderItem.network)
         }
