+++ app/src/main/java/com/tari/android/wallet/ui/fragment/settings/torBridges/customBridges/CustomTorBridgesViewModel.kt
@@ -69,6 +69,6 @@ class CustomTorBridgesViewModel : CommonViewModel() {
             resourceManager.getString(R.string.common_error_title),
             resourceManager.getString(R.string.tor_bridges_incorrect_format)
         )
-        _modularDialog.postValue(args.getModular(resourceManager))
+        modularDialog.postValue(args.getModular(resourceManager))
     }
 }
