+++ app/src/main/res/layout/dialog_module_icon.xml
@@ -6,14 +6,15 @@
     android:layout_height="wrap_content"
     android:clipToPadding="false"
     android:paddingBottom="20dp"
+    android:paddingTop="16dp"
     android:clipChildren="false"
     android:orientation="vertical">
     
     <com.tari.android.wallet.ui.component.tari.TariPrimaryBackground
         android:layout_width="80dp"
         android:layout_height="80dp"
-        android:elevation="13dp"
-        app:elevation="13dp"
+        android:elevation="8dp"
+        app:elevation="8dp"
         android:layout_gravity="center_horizontal"
         app:cornerRadius="40dp">
